PR-SUG: move the message pump assertion functionality abstracter
